Moving all release related steps to release.py
